Folded the highly-redundant matrix.Assembler class into _Slice
