+++ src/cr/cube/matrix/__init__.py
@@ -1,5 +1,3 @@
 # encoding: utf-8
 
 """A matrix is the 2D cube-data partition used by a slice."""
-
-from cr.cube.matrix.assembler import Assembler  # noqa: F401
